95% Parser Done & Fixed Literal Bug
Parser is 95% done, Still needs Expression to be implemented when display() is run. 

Also, When Parser is run it relies on TokenStream.java to assign the types. When we implemented boolean types, we changed the type "Literal" to "Integer-Literal" and "Boolean-Literal". ConcreteSyntax.java does not recognize these new types and throws a fit. If you want we can rework ConcreteSyntax.java to recognize "Integer-Literal" and "Boolean-Literal".